添加上拉刷新

+++ iOS-git-osc/Project.h
@@ -12,7 +12,7 @@
 
 @interface Project : NSObject
 
-+ (NSArray *)loadExtraProjectType:(int)type onPage:(int)page;
++ (NSArray *)loadExtraProjectType:(NSInteger)type onPage:(int)page;
 + (NSArray *)getProjectTreeWithID:(int64_t)projectID Branch:(NSString *)branch Path:(NSString *)path;
 + (NSString *)getFileContent:(int64_t)projectID Path:(NSString *)path Branch:(NSString *)branch;
 + (NSArray *)getOwnProjectsOnPage:(int)page;
